Add terminate statements, create Kernel object on CLWrapper
I have no idea if calling terminate in this fashion is bad, however I decided to use it because implementing error handling would add a lot of complexity to the project. I've also looked up if the resources are freed upon termination and it seems as though there is no need to manually free up everything (again, adding complexity) before terminating the program. I think for the scope of this project it should be ok.